// backend/src/common/enums/estado-declaracion-carga.enum.ts
export enum EstadoDeclaracionCarga {
  NO_INICIADO = "NO_INICIADO",
  BORRADOR = "BORRADOR",
  PENDIENTE_ENVIO = "PENDIENTE_ENVIO",
  ENVIADO_DOCENTE = "ENVIADO_DOCENTE",
  OBSERVADO_DPTO = "OBSERVADO_DPTO",
  SUBSANADO = "SUBSANADO",
  VALIDADO_DPTO = "VALIDADO_DPTO",
  OBSERVADO_FACULTAD = "OBSERVADO_FACULTAD",
  APROBADO_FACULTAD = "APROBADO_FACULTAD",
  CERRADO = "CERRADO",
  ANULADO = "ANULADO",
}